Rename file name Braille.ts to BrailleCharacter.ts
