+++ src/models/BrailleCharacter.ts
